Update index.html
There are some style improvements, that should be moved to a different CSS file. For example,, it should be created a "style-overrides.css" file, which can be referenced in a link tag like this: 

<link rel = "stylesheet" href="./css/styles-overriden.css">

First time I had some JavaScript functions, that was changing the styles. Now almost everything is changed from the CSS attributes, Maybe this is a more efficient way, since a JavaScript method was querying some HTML elements, and can slow down the page loading.